8.1 video

+++ controllers/videoController.js
@@ -55,9 +55,11 @@ export const postUpload = async (req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
-    console.log(newVideo);
+    req.user.videos.push(newVideo);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -68,7 +70,7 @@ export const videoDetail = async (req, res) => {
         }
     } = req;
     try {
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate('creator');
         res.render("videoDetail", {
             pageTitle: video.title,
             video
@@ -86,10 +88,15 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", {
-            pageTitle: `Edit ${video.title}`,
-            video
-        });
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", {
+                pageTitle: `Edit ${video.title}`,
+                video
+            });
+        }
+
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -124,9 +131,15 @@ export const deleteVideo = async (req, res) => {
         }
     } = req;
     try {
-        await Video.findOneAndRemove({
-            _id: id,
-        });
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({
+                _id: id,
+            });
+        }
+
     } catch (error) {
         console.log(error);
     };
